requirements.txt
# Core dependencies
streamlit==1.28.0
torch==2.0.0
transformers==4.30.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Data processing
numpy==1.24.0
pandas==2.0.0

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.14.0

# PDF generation
reportlab==4.0.0

# Utilities
python-dateutil==2.8.2
